requirements.txt
aiohttp
aiohttp-jinja2
telethon
cryptg
pillow
setuptools
aiohttp_session[secure]
python-dotenv
requests
